backend/requirements.txt
fastapi
uvicorn[standard]
motor
pymongo[srv]
redis
python-dotenv
pydantic
pandas
numpy
